backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.5.0
scipy>=1.14.0
statsmodels>=0.14.0
textblob>=0.18.0
sqlparse>=0.5.0
pydantic-settings>=2.6.0
redis>=5.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
